+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # Version 0.0.1
-* First publish
+
+- First publish
 
 ---
 
 # Version 0.0.2
-* Bug fixes: import issue
+
+- Bug fixes: import issue
 
 ---
 
 # Version 0.0.3
-* Added complete API in `README.md`
+
+- Added complete API in `README.md`
+
+--
+
+# Version 0.0.4
+
+- Added new features: `lower`, `upper`, and `ip`